Cleaner .h/.cpp, overloaded for WiFi parameters
A more pure .h/.cpp separation - implementation in the .cpp file
and the defines in the .h file.

Added also now the option to include WiFi SSID, Password as params
via an overloaded function. There is now logic to deduce if the
params were given and they were - those will be used.

If the paramers were note given (and we're using WiFi) we will
revert using the params given in the `mbed_app.json` -file.

That should sort out issue https://github.com/ARMmbed/easy-connect/issues/56.

Rebased on top of the no-password commit by Andrew Chong, however
I think the the comparison should >1 rather than >2 - the password
could be just one character as well.
